# Stash.Tests/Registry/OpenApi/Snapshots/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Stash Package Registry",
    "description": "REST API for the Stash language\u0027s package registry \u2014 packages, scopes, organizations, search, auth.",
    "contact": {
      "name": "Stash project",
      "url": "https://github.com/cmoraru/stash-lang"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://registry.invalid"
    }
  ],
  "paths": {
...
    "/api/v1/scopes/{scope}": {
      "get": {
        "tags": [
          "Scopes"
        ],
        "operationId": "Scopes_GetScope",
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Scopes"
        ],
        "operationId": "Scopes_DeleteScope",
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scopes": {
      "post": {
        "tags": [
          "Scopes"
        ],
        "operationId": "Scopes_ClaimScope",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimScopeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimScopeRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/ClaimScopeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scopes/{scope}/verify": {
      "post": {
        "tags": [
          "Scopes"
        ],
        "operationId": "Scopes_VerifyScope",
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeVerifyResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClaimScopeRequest": {
        "required": [
          "scope",
          "owner"
        ],
        "type": "object",
        "properties": {
          "scope": {
            "type": [
              "null",
              "string"
            ]
          },
          "owner_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ScopeOwnerTypes"
              }
            ]
          },
          "owner": {
            "type": [
              "null",
              "string"
            ]
          },
          "verification_method": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ScopeChallengeBody": {
        "required": [
          "method",
          "record_name",
          "record_value",
          "expires_at"
        ],
        "type": "object",
        "properties": {
          "method": {
            "type": "string"
          },
          "record_name": {
            "type": "string"
          },
          "record_value": {
            "type": "string"
          },
          "expires_at": {
            "type": "string"
          }
        }
      },
      "ScopeDetailResponse": {
        "required": [
          "scope",
          "owner_type"
        ],
        "type": "object",
        "properties": {
          "scope": {
            "type": "string"
          },
          "owner_type": {
            "$ref": "#/components/schemas/ScopeOwnerTypes"
          },
          "owner": {
            "type": [
              "null",
              "string"
            ]
          },
          "state": {
            "type": [
              "null",
              "string"
            ]
          },
          "challenge": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ScopeChallengeBody"
              }
            ]
          }
        }
      },
      "ScopeOwnerTypes": {
        "enum": [
          "user",
          "org",
          "system"
        ]
      },
      "ScopeVerifyResponse": {
        "required": [
          "scope",
          "method",
          "message"
        ],
        "type": "object",
        "properties": {
          "scope": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}